requirements.txt
aiohttp==3.6.2
python-snappy==0.5.4
pytest==5.4.2
pytest-asyncio==0.12.0
